Split up ScalingPolicy to simplify model-based scaling
Now ScalingPolicy is abstract, the existing use-cases are grouped under
ReactiveScalingPolicy while the model-based scaling happens using the
new ModelBasedScalingPolicy entity. For this, only a model is given as
child which then contains both the reward and the stimuli used as input.

The Transformations are changed as well in order to be able to extract
the adjustment magnitude from the new ModelBasedScalingPolicy.